+++ lib/handlers/listing.js
@@ -1,6 +1,10 @@
 // Service Wrappers
 var geocoder = require('../wrappers/geocoder.js');
 
+// Other Dependencies
+var Consts = require('../consts.js');
+
+// Listing Model Packages
 var listing_model_path = '../models/listings/';
 function pt(f) {
     return listing_model_path + f;
@@ -33,73 +37,31 @@ exports.getListingsByUserID = getListingsByUserID;
 function makeNewListingFromPost(body, user_id, callback) {
     body.user_id = user_id;
     var paths = [];
-    if ('type' in body) {
-        if (body.type === misc_listing.name) {
-            var miscListing = new Listings.MiscListing;
-            for (var path in Listings._MiscSchema.paths) {
-                var first_period = path.indexOf('.');
-                var real_path = path;
-                if (first_period != -1) {
-                    real_path = path.substring(0, first_period);
-                }
-                if (paths.indexOf(real_path) == -1) {
-                    paths.push(real_path);
-                }
+    if ('type' in body and body.type in listing_name_map) {
+        var sel_listing = listing_name_map[body.type];
+        var listing = new sel_listing.model;
+        var listing_schema = sel_listing.schema;
+        for (var path in listing_schema.paths) {
+            var first_period = path.indexOf('.');
+            var real_path = path;
+            if (first_period != -1) {
+                real_path = path.substring(0, first_period);
             }
-            Utils.copyObjectData(miscListing, body, paths);
-            return updateMiscListingLatLng(miscListing, function(result) {
-                if (result) {
-                    return callback({
-                        type: body.type,
-                        listing: miscListing
-                    });
-                } else {
-                    return callback(false);
-                }
-            });
-        } else if (body.type === sublet_listing.name) {
-            var subletListing = new Listings.SubletListing;
-            for (var path in Listings._SubletSchema.paths) {
-                var first_period = path.indexOf('.');
-                var real_path = path;
-                if (first_period != -1) {
-                    real_path = path.substring(0, first_period);
-                }
-                if (paths.indexOf(real_path) == -1) {
-                    paths.push(real_path);
-                }
+            if (paths.indexOf(real_path) == -1) {
+                paths.push(real_path);
             }
-            Utils.copyObjectData(subletListing, body, paths);
-            return updateSubletListingLatLng(subletListing, function(result) {
-                if (result) {
-                    callback({
-                        type: body.type,
-                        listing: subletListing
-                    });
-                } else {
-                    return callback(false);
-                }
-            });
-        } else if (body.type === sublet_listing.name) {
-            var bookListing = new Listings.BookListing;
-            for (var path in Listings._BookSchema.paths) {
-                var first_period = path.indexOf('.');
-                var real_path = path;
-                if (first_period != -1) {
-                    real_path = path.substring(0, first_period);
-                }
-                if (paths.indexOf(real_path) == -1) {
-                    paths.push(real_path);
-                }
-            }
-            Utils.copyObjectData(bookListing, body, paths);
-            return callback({
-                type: body.type,
-                listing: bookListing
-            });
-        } else if (book.type === book_listing.name) {
-
         }
+        Utils.copyObjectData(listing, body, paths);
+        return updateListingLatLng(listing, function(result) {
+            if (result) {
+                return callback({
+                    type: body.type,
+                    listing: listing
+                });
+            } else {
+                return callback(false);
+            }
+        });
     }
     return callback(false);
 }
@@ -156,6 +118,32 @@ function updateMiscListingLatLng(listing, callback) {
             }
         });
     }
+}
+
+function secureListings(request, listings, include_id) {
+    return listings.map(function(e) {
+        return secureListing(request, e, include_id);
+    }).filter(function(e) {
+        //
+    });
+}
+
+function secureListing(request, listing, include_id) {
+    if ("toSecure" in listing && (typeof(listing.toSecure) === typeof(secureListing))) {
+        if (Common.ensureLoginSession(request)) {
+            return listing.toSecure(request.session.login.who.role, request.session.login.who.id, include_id);
+        } else
+            return listing.toSecure(Consts.ROLE_INVALID, include_id);
+    } else if (Common.ensureLoginSession(request)) {
+        //TODO: warn and log this case... should NEVER happen!
+        if (Consts.checkPriv(request.session.who.role, Consts.ROLE_ADMIN)) {
+            return listing;
+        } else if (Consts.checkPriv(request.session.who.role, Consts.ROLE_MODERATOR_PUBLIC)
+            && request.session.who.role & Consts.extractMask(listing.role_mask)) {
+            return listing;
+        }
+    }
+    return {};
 }
