+++ lib/route/listing.js
@@ -103,33 +103,6 @@ exports.postListingById = function(request, response) {
         }
     });
 }
-
-function secureListings(request, listings, include_id) {
-    return listings.map(function(e) {
-        return secureListing(request, e, include_id);
-    }).filter(function(e) {
-        //
-    });
-}
-
-function secureListing(request, listing, include_id) {
-    if ("toSecure" in listing && (typeof(listing.toSecure) === typeof(secureListing))) {
-        if (Common.ensureLoginSession(request)) {
-            return listing.toSecure(request.session.login.who.role, request.session.login.who.id, include_id);
-        } else
-            return listing.toSecure(Consts.ROLE_INVALID, include_id);
-    } else if (Common.ensureLoginSession(request)) {
-        //TODO: warn and log this case... should NEVER happen!
-        if (Consts.checkPriv(request.session.who.role, Consts.ROLE_ADMIN)) {
-            return listing;
-        } else if (Consts.checkPriv(request.session.who.role, Consts.ROLE_MODERATOR_PUBLIC)
-            && request.session.who.role & Consts.extractMask(listing.role_mask)) {
-            return listing;
-        }
-    }
-    return {};
-}
-
 exports.getListingByUserId = function(request, response) {
     response.set('Content-Type', 'application/json');
